Moving features around no longer causes a rave party
Closes #841

Not exactly sure why, but it would start sending duplicate updateFeature calls, and then doubling the number, and doubling again in certain hard-to-reproduce circumstances.

The root cause seems to have been prepareFeaturesForMap() being called after the update and re-updating the map with the new data.

Our fix is to...just not do that and to leave it at updating the Redux store with the new position of the feature. It was redundant to prepareFeaturesForMap() anyway, as the map is updated already by the Modify Interaction that causes all of this in the first place!

We'd also see weird artefacts like the leftover modify markers/grab points on the map at the positions it used to be in.

We were seeing multiple updateFeature() calls because the interaction was actually returning multiple instances of the feature object as a result of the user moving it (all with the same feature id).

It's probably some subtle OL bug/misunderstanding in my part where you can't change the underlying feature data while a Modify Interaction is active - at least not without refreshing the source behind it somehow.

Anywho, this is a good enough fix.